Define a simple text input protocol #8
