+++ src/bootx64/src/uefi/tables/system.rs
@@ -19,6 +19,6 @@ pub struct System<'a> {
     firmware_vendor: char16::String,
     firmware_revision: u32,
     console_in_handle: handle::Handle<'a>,
-    con_in: &'a simple_text_input::SimpleTextInput,
+    con_in: &'a simple_text_input::SimpleTextInput<'a>,
 }
