+++ src/bootx64/src/uefi/types/event.rs
@@ -1,10 +1,8 @@
-// Reference
+use super::void;
+
+// References
 // https://uefi.org/sites/default/files/resources/UEFI_Spec_2_9_2021_03_18.pdf
 // 2.3.1 Data Types
-
-pub mod char16;
-pub mod event;
-pub mod handle;
-pub mod status;
-pub mod void;
+// EFI_EVENT
+pub type Event<'a> = &'a void::Void;
