+++ src/bootx64/src/uefi/types/status.rs
@@ -2,5 +2,6 @@
 // https://uefi.org/sites/default/files/resources/UEFI_Spec_2_9_2021_03_18.pdf
 // 2.3.1 Data Types
 // Appendix D
+// EFI_STATUS
 pub type Status = u64;
